+++ eNear/scripts/1_deploy_eNear.js
@@ -1,13 +1,31 @@
+const prompt = require('prompt-sync')()
+
 async function main() {
   const [deployer] = await ethers.getSigners()
   const deployerAddress = await deployer.getAddress()
   console.log(
-    "Deploying eNear logic contract with the account:",
+    "Deploying eNear contract with the account:",
     deployerAddress
   )
 
+  const tokenName = prompt('Token name? ');
+  const tokenSymbol = prompt('Token symbol? ');
+  const nearConnector = prompt('Near connector account? ');
+  const proverAddress = prompt('Prover address? ');
+  const minBlockAcceptanceHeight = prompt('Min block height? ');
+  const adminAddress = prompt('eNear and proxy admin? ');
+  const pausedFlagValues = prompt('Paused flag values? ');
+
   const eNearFactory = await ethers.getContractFactory("eNear")
-  const eNear = await eNearFactory.deploy()
+  const eNear = await eNearFactory.deploy(
+    tokenName,
+    tokenSymbol,
+    Buffer.from(nearConnector, 'utf-8'),
+    proverAddress,
+    minBlockAcceptanceHeight,
+    adminAddress,
+    pausedFlagValues
+  )
 
   await eNear.deployed()
