+++ eNear/scripts/2_pause_e_near_method.js
